`splits.txt`: Mark absolute Windows-style paths as invalid
May be of mild utility for those using a symbol map as a starting point, for which paths like these are emitted for MetroTRK's `__exception.s` and the like.

Only matches paths that start with `[[:alpha:]]:\\`, but these are probably the only ones that naturally occur anyways, so who cares.